# Day_35_hardened_notes_api/internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package db

import "time"

type Note struct {
	ID        int64
	UserID    int64
	Title     string
	Body      string
	CreatedAt time.Time
	UpdatedAt time.Time
}

type User struct {
	ID           int64
	Email        string
	PasswordHash string
	CreatedAt    time.Time
	UpdatedAt    time.Time
}