# Day_35_hardened_notes_api/internal/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Notes API
  version: "1.0.0"
  description: |
    Personal notes API — multi-user, JWT-authenticated, cursor-paginated,
    rate-limited, idempotent on retries. Week 5 ship of a 15-week Go
    backend learning plan.

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - { name: health, description: "Service health" }
  - { name: auth,   description: "Registration, login, token refresh" }
  - { name: notes,  description: "Per-user notes CRUD" }
  - { name: audit,  description: "User-scoped audit log" }

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "`Authorization: Bearer <access_token>` from /auth/login"

  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:    { type: string, example: NOT_FOUND }
            message: { type: string, example: "note not found" }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string, example: email }
                  issue: { type: string, example: "must be a valid email" }

    User:
      type: object
      required: [id, email, created_at, updated_at]
      properties:
        id:         { type: integer, format: int64, example: 1 }
        email:      { type: string, format: email, example: alice@example.com }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:    { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 72 }

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:    { type: string, format: email }
        password: { type: string }

    LoginResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:  { type: string }
        refresh_token: { type: string }
        token_type:    { type: string, example: Bearer }
        expires_in:    { type: integer, example: 900 }

    RefreshRequest:
      type: object
      required: [refresh_token]
      properties: { refresh_token: { type: string } }

    LogoutRequest:
      type: object
      required: [refresh_token]
      properties: { refresh_token: { type: string } }

    Note:
      type: object
      required: [id, user_id, title, body, created_at, updated_at]
      properties:
        id:         { type: integer, format: int64 }
        user_id:    { type: integer, format: int64 }
        title:      { type: string }
        body:       { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    CreateNoteRequest:
      type: object
      required: [title]
      properties:
        title: { type: string, maxLength: 200 }
        body:  { type: string, maxLength: 10000 }

    UpdateNoteRequest:
      type: object
      required: [title, body]
      properties:
        title: { type: string, maxLength: 200 }
        body:  { type: string, maxLength: 10000 }

    PatchNoteRequest:
      type: object
      description: "At least one of title/body must be present."
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        body:  { type: string, maxLength: 10000 }

    NoteListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Note' }
        next_cursor:
          type: string
          description: "Opaque cursor; pass as ?after=… for the next page."

    NoteRef:
      type: object
      properties:
        id:    { type: integer, format: int64 }
        title: { type: string }

    AuditEntry:
      type: object
      required: [id, user_id, action, target_type, target_id, created_at]
      properties:
        id:          { type: integer, format: int64 }
        user_id:     { type: integer, format: int64 }
        action:      { type: string, enum: [note.created, note.updated, note.patched, note.deleted] }
        target_type: { type: string, example: note }
        target_id:   { type: integer, format: int64 }
        metadata:    { type: object, additionalProperties: true, nullable: true }
        created_at:  { type: string, format: date-time }

    AuditEntryWithNote:
      allOf:
        - $ref: '#/components/schemas/AuditEntry'
        - type: object
          properties:
            note:
              nullable: true
              allOf:
                - $ref: '#/components/schemas/NoteRef'

    AuditListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/AuditEntry' }

    AuditListWithNotesResponse:
      type: object
      required: [data, strategy]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/AuditEntryWithNote' }
        strategy: { type: string, enum: [join, in_batch, naive] }

  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema: { type: string, maxLength: 255 }
      description: |
        Optional per-attempt unique key for safe retries. Same key + same body
        replays the original response. Same key + different body returns 422.

  responses:
    BadRequest:
      description: Malformed request
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: Missing or invalid auth
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: Conflicting state
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Validation:
      description: Request validation failed
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After: { schema: { type: integer }, description: Seconds until unblocked. }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }

paths:
  /healthz:
    get:
      tags: [health]
      summary: Liveness probe
      security: []
      responses:
        '200':
          description: Service is up
          content:
            application/json:
              schema: { type: object, properties: { status: { type: string, example: ok } } }

  /auth/register:
    post:
      tags: [auth]
      summary: Register a new user
      security: []
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/RegisterRequest' } } } }
      responses:
        '201':
          description: User created
          headers:
            Location: { schema: { type: string } }
          content: { application/json: { schema: { $ref: '#/components/schemas/User' } } }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Validation' }
        '429': { $ref: '#/components/responses/TooManyRequests' }

  /auth/login:
    post:
      tags: [auth]
      summary: Login and receive access + refresh tokens
      security: []
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } } } }
      responses:
        '200':
          description: Authentication successful
          content: { application/json: { schema: { $ref: '#/components/schemas/LoginResponse' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/Validation' }
        '429': { $ref: '#/components/responses/TooManyRequests' }

  /auth/refresh:
    post:
      tags: [auth]
      summary: Exchange a refresh token for a new pair (rotates the refresh)
      security: []
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/RefreshRequest' } } } }
      responses:
        '200':
          description: New token pair issued
          content: { application/json: { schema: { $ref: '#/components/schemas/LoginResponse' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /auth/logout:
    post:
      tags: [auth]
      summary: Revoke a refresh token
      security: []
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/LogoutRequest' } } } }
      responses:
        '204': { description: Logged out }

  /auth/me:
    get:
      tags: [auth]
      summary: Return the authenticated user
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: The current user
          content: { application/json: { schema: { $ref: '#/components/schemas/User' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /notes:
    get:
      tags: [notes]
      summary: List notes for the authenticated user (cursor paginated)
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: limit,  in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: after,  in: query, schema: { type: string } }
        - { name: sort,   in: query, schema: { type: string, enum: [asc, desc], default: desc } }
      responses:
        '200':
          description: List of notes
          content: { application/json: { schema: { $ref: '#/components/schemas/NoteListResponse' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    post:
      tags: [notes]
      summary: Create a note
      security: [{ BearerAuth: [] }]
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateNoteRequest' } } } }
      responses:
        '201':
          description: Note created
          headers:
            Location: { schema: { type: string } }
            Idempotent-Replayed: { schema: { type: string, enum: ["true"] }, description: Present when this is a replay. }
          content: { application/json: { schema: { $ref: '#/components/schemas/Note' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { description: Idempotency in-flight }
        '422': { $ref: '#/components/responses/Validation' }
        '429': { $ref: '#/components/responses/TooManyRequests' }

  /notes/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64, minimum: 1 } }
    get:
      tags: [notes]
      summary: Get a note by id
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: The note
          content: { application/json: { schema: { $ref: '#/components/schemas/Note' } } }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [notes]
      summary: Replace a note's title + body
      security: [{ BearerAuth: [] }]
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateNoteRequest' } } } }
      responses:
        '200':
          description: The updated note
          content: { application/json: { schema: { $ref: '#/components/schemas/Note' } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Validation' }
    patch:
      tags: [notes]
      summary: Patch a note (at least one field)
      security: [{ BearerAuth: [] }]
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PatchNoteRequest' } } } }
      responses:
        '200':
          description: The patched note
          content: { application/json: { schema: { $ref: '#/components/schemas/Note' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Validation' }
    delete:
      tags: [notes]
      summary: Delete a note
      security: [{ BearerAuth: [] }]
      responses:
        '204': { description: Note deleted }
        '404': { $ref: '#/components/responses/NotFound' }

  /audit:
    get:
      tags: [audit]
      summary: List recent audit entries for the authenticated user
      description: |
        Without `include=notes`, returns plain audit entries. With it, joins
        the referenced note via one of three strategies (see Day 30 N+1).
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: limit,    in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
        - { name: include,  in: query, schema: { type: string, enum: [notes] } }
        - { name: strategy, in: query, schema: { type: string, enum: [join, in_batch, naive], default: join } }
      responses:
        '200':
          description: Audit entries
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/AuditListResponse'
                  - $ref: '#/components/schemas/AuditListWithNotesResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }